Update some dependencies
The main impetus for this PR is to update the very outdated rack dep, which keeps registering critical issues in our vulnerability scanners.

While I was at it, I updated a few others. Would be nice to update rack-test to 2.1 and mocha as well, but they require more extensive changes.