fix(deps): Add faraday-multipart as actual dep
We ran into an issue recently where we changed our code to not load faraday-multipart on startup and hit an issue with the looker-sdk complaining it could not find `Faraday::Multipart`.
Since that code is used at runtime, that should be an actual dependency and not a dev dep only.